Fix jUnpack, give jCopula a new hi def, and more
jUnpack is no longer a region, now it's just a match def.
This fixes an issue where lines containing jQuote before jUnpack would
highlight incorrectly. Found while golfing.

This also required the creation of a new jQuote match def that is only
contained in jUnpack so that the quotes highlight as PreProc.

I've also moved jCopula to a highlight definition. Before it was
PreProc, now it's Structure. No reason other than nothing else was using
it, and most colorscheme's highlight it the same as PreProc anyway,
but... it's now got it's own def for those wishing to customise.

Lastly, jIdentifier can contain underscores too. This almost a
non-issue, as jIdentifier has no coloring (like unmactched syntax.)
Again this is to ensure they are matched in case of user customisation.